Deleted tests
One test fails for reasons that I don't understand. I think it is
probably a test that gets autogenerated when you create a Spring project
from the Spring web page.

The other test requires a running Blaze store to work, it is never going to
work with a standalone feedback agent build.